Moved generic APIs to main package. Consistent annotation names.
