+++ implementation/src/main/java/io/smallrye/config/KeyMap.java
@@ -1,4 +1,4 @@
-package io.smallrye.config.mapper;
+package io.smallrye.config;
 
 import java.io.Serializable;
 import java.util.HashMap;
