+++ implementation/src/main/java/io/smallrye/config/mapper/WithDefault.java
@@ -12,6 +12,6 @@ import java.lang.annotation.Target;
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Default {
+public @interface WithDefault {
     String value();
 }
